Finished
Finished website. Gave a section the same background color as the body so that the color would reach all the way to the bottom. Fixed the footer so that after the section element it has a container collapse. Fixed a border for one of the images, corrected HTML code so that it got the border it needed. Finalized it all with finishing comments throughout the HTML and what I added to the CSS.